Env

Description

This library is meant to be used with the basic environment variable support that Haskell has (System.Environment).

import the library (import Env) and System.Environment (we need getEnv)

import Env
import System.Environment

use the loadEnv or loadEnvFile function

loadEnv
loadEnvFile "pathToEnv.env"

to get your values just use the getEnv "nameOfKey" from System.Environment

inside the env file -> HELLO=WORLD

getEnv "HELLO"
"WORLD"

The difference between loadEnv and loadEnvFile.

loadEnv is meant to search automatically for a .env file. It will throw an error if 2 or more (or none) .env files are found.

loadEnvFile is meant to be used with the path to the .env file (much safer).

Documentation

loadEnv :: IO ()

Load the first env file found

loadEnvFile :: String -> IO ()

Load the env file specified on the path